BAU: Option to deploy to sandpit without the gds cli
Not all devs are setup to user the gds cli, and we are trying to move everyone over to AWS SSO.

- Remove the dependency on the gds cli from the script (but suggest how to use it if required).
- Sandpit deploy is split into two separate parts which need to be run independently in order to make use of two separate AWS accounts.
1. Build and push the docker image to digital-identity-tools-dev, options -i and -b.
2. Deploy the image and app by running the terraform in gds-di-development, default option.

+++ deploy-sandpit.sh
@@ -25,16 +25,19 @@ USAGE
 }
 
 BUILD=0
+IMAGE=0
 TERRAFORM=0
 TERRAFORM_OPTS="-auto-approve"
 if [[ $# == 0 ]]; then
-  BUILD=1
   TERRAFORM=1
 fi
 while [[ $# -gt 0 ]]; do
   case $1 in
   -b | --build)
     BUILD=1
+    ;;
+  -i | --image)
+    IMAGE=1
     ;;
   -t | --terraform)
     # shellcheck disable=SC2034
@@ -54,35 +57,49 @@ while [[ $# -gt 0 ]]; do
   shift
 done
 
-echo "Generating temporary ECR credentials..."
-gds aws di-tools-dev -- aws ecr get-login-password --region eu-west-2 | docker login --username AWS --password-stdin "${REPO_URL}"
+if [[ $IMAGE == "1" ]]; then
+  if [[ -z "${AWS_ACCESS_KEY_ID:-}" || -z "${AWS_SECRET_ACCESS_KEY:-}" ]]; then
+    echo "!! AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY must be set in the environment." >&2
+    echo "!! Use AWS SSO digital-identity-tools-dev or for gds-cli: gds aws di-tools-dev -- ${0}" >&2
+    exit 1
+  fi
 
-if [[ $BUILD == "1" ]]; then
-  echo "Building image..."
-  docker buildx build --platform=linux/amd64 -t "${REPO_NAME}" .
-  echo "Tagging image..."
-  docker tag "${REPO_NAME}:latest" "${REPO_URL}:${IMAGE_TAG}"
+  echo "Generating temporary ECR credentials..."
+  aws ecr get-login-password --region eu-west-2 | docker login --username AWS --password-stdin "${REPO_URL}"
 
-  echo "Pushing image..."
-  docker push "${REPO_URL}:${IMAGE_TAG}"
-  IMAGE_DIGEST="$(docker inspect "${REPO_URL}:${IMAGE_TAG}" | jq -r '.[0].RepoDigests[0] | split("@") | .[1]')"
-  echo "Digest = ${IMAGE_DIGEST}"
-  echo "Complete"
-else
-  docker pull "${REPO_URL}:${IMAGE_TAG}"
-  IMAGE_DIGEST="$(docker inspect "${REPO_URL}:${IMAGE_TAG}" | jq -r '.[0].RepoDigests[0] | split("@") | .[1]')"
+  if [[ $BUILD == "1" ]]; then
+    echo "Building image..."
+    docker buildx build --platform=linux/amd64 -t "${REPO_NAME}" .
+    echo "Tagging image..."
+    docker tag "${REPO_NAME}:latest" "${REPO_URL}:${IMAGE_TAG}"
+
+    echo "Pushing image..."
+    docker push "${REPO_URL}:${IMAGE_TAG}"
+    IMAGE_DIGEST="$(docker inspect "${REPO_URL}:${IMAGE_TAG}" | jq -r '.[0].RepoDigests[0] | split("@") | .[1]')"
+    echo "Digest = ${IMAGE_DIGEST}"
+    echo "Complete"
+  else
+    docker pull "${REPO_URL}:${IMAGE_TAG}"
+    IMAGE_DIGEST="$(docker inspect "${REPO_URL}:${IMAGE_TAG}" | jq -r '.[0].RepoDigests[0] | split("@") | .[1]')"
+  fi 
+  export IMAGE_DIGEST_EXT=$IMAGE_DIGEST
+  echo "bob $IMAGE_DIGEST_EXT"
+  exit 0
 fi
 
 if [[ $TERRAFORM == "1" ]]; then
   echo -n "Getting AWS credentials ... "
-  eval "$(gds aws digital-identity-dev -e)"
-  echo "done!"
-
+  if [[ -z "${AWS_ACCESS_KEY_ID:-}" || -z "${AWS_SECRET_ACCESS_KEY:-}" ]]; then
+    echo "!! AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY must be set in the environment." >&2
+    echo "!! Use AWS SSO digital-identity-dev or for gds-cli: gds aws digital-identity-dev -- ${0}" >&2
+    exit 1
+  fi
+  
   echo "Running Terraform..."
   pushd "${DIR}/ci/terraform" >/dev/null
   rm -rf .terraform/
   terraform init -backend-config=sandpit.hcl
-  terraform apply ${TERRAFORM_OPTS} -var-file sandpit.tfvars -var "image_uri=${REPO_URL}" -var "image_digest=${IMAGE_DIGEST}"
+  terraform apply ${TERRAFORM_OPTS} -var-file sandpit.tfvars -var "image_uri=${REPO_URL}" -var "image_digest=${IMAGE_DIGEST_EXT}"
 
   if [[ $TERRAFORM_OPTS != "-destroy" ]]; then
     echo -n "Waiting for ECS deployment to complete ... "
